I added a sample "HTA" file for running Selenium, and added a sample test suite and test case that visits google.com, performs a search, and verifies results.
If you tried to run the Google test case from a "regular" Selenium setup running on your server, the test would fail with a message saying "Permission Denied."  Cross-site scripting security in the web browser prevents the JavaScript that powers Selenium running on your server from reading or editing the contents of the page coming from google.com. 

One way to get around the cross site scripting problem is to run the Selenium Test Runner as an (HyperText Application) "HTA" application, which removes security restrictions that normal web pages are subjected to. 

Running the test runner as an HTA, the google test suite should pass successfully.

The up-side to this approach is that we solve the cross-site scripting problem. It also solves the problem of having to install Selenium on the same server that you want to test. The down-side is that this works with Internet Explorer only. 



git-svn-id: https://svn.openqa.org/svn/selenium/trunk@249 0891141a-5dea-0310-ad27-ebc607f31677