OK, I didn't hear from anyone, so I'm just going to check this in and see if the build works :)
This is my ghetto/temp solution for solving the 2-connections-per-host problem when running PI mode. The issue is that if there are many frames all waiting on the server with an open socket, the 2 connection limit can deadlock the test because another window that needs to call back to the server cannot.

There are two ways to fix this:

1) Instead of blocking threads indefinitely, kick the thread back with a notification that the socket request should retry again later. This lets stuck threads "get through". 

2) Implement a cross-domain compatible alternative to XmlHttpRequest (using iframes and/or script blocks).

This checkin is an attempt at technique #1. It breaks timeouts (they don't happen anymore), so it isn't complete, but I want to see how the CI build takes to this significant change.

This implementation is also very ghetto. On the other hand, the rest of the code is pretty ghetto too, so I guess it's not a big deal.

git-svn-id: https://svn.openqa.org/svn/selenium/trunk@1765 0891141a-5dea-0310-ad27-ebc607f31677